fix(Sidebar): due dates were buggy (NaN) and not displaying
