+++ frontend/src/layouts/SidebarLayout/RequestFeed/RequestFeedGroups/RequestItem.tsx
@@ -18,6 +18,7 @@ import { theme } from "~ui/theme";
 import { RequestContentSnippet } from "../RequestContentSnippet";
 import { RequestMessagePreview } from "../RequestMessagePreview";
 import { RequestParticipants } from "../RequestParticipants";
+import { getNearestTaskDueDateForCurrentUser } from "./shared";
 
 interface Props {
   topic: TopicEntity;
@@ -88,19 +89,18 @@ export const RequestItem = observer(function RequestItem({ topic }: Props) {
 });
 
 const OpenTopicSubtitle = observer(function OpenTopicSubtitle({ topic }: { topic: TopicEntity }) {
+  const currentUserUnfinishedDueDate = getNearestTaskDueDateForCurrentUser(topic);
+
   return (
     <>
       {/* Content snippet requires booting up rich editor with plugins, lets make it lazy so it renders in next 'tick' */}
-      {!topic.currentUserUnfinishedTaskWithEarliestDueDate && (
+      {!currentUserUnfinishedDueDate && (
         <LazyRender fallback={<div>&nbsp;</div>}>
           <RequestContentSnippet topic={topic} />
         </LazyRender>
       )}
 
-      {topic.currentUserUnfinishedTaskWithEarliestDueDate && (
-        // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-        <>{getRelativeDueTimeLabel(topic.currentUserUnfinishedTaskWithEarliestDueDate.dueDate!)}</>
-      )}
+      {currentUserUnfinishedDueDate && <>{getRelativeDueTimeLabel(currentUserUnfinishedDueDate)}</>}
     </>
   );
 });
